Use create_densities within perturbation_auf instead of calculate_local_density, as a better, more flexible implementation.
